1. Added print_error_log, so all MISSING_PROPERTY errors print only in log, and MISSING_FILE goes to console. This is because MISSING_PROPERTY does not fail the validation. 2. Added check before installation of missing py_modules. 3. Added ant_tasks.xml to run 'getRemoteFile' task

+++ scripts/setup.sh
@@ -4,7 +4,11 @@
 # installs python module dependencies
 
 install() {
-    pushd . &&  if [ -d "temp" ]; then rm -Rf temp; fi && mkdir temp && tar zxf $1 -C temp && cd temp && cd * && python setup.py install --user && popd && rm -rf temp
+    module=$2
+    python -c "import $module"
+    if [ $? == 1 ]; then
+        pushd . &&  if [ -d "temp" ]; then rm -Rf temp; fi && mkdir temp && tar zxf $1 -C temp && cd temp && cd * && python setup.py install --user && popd && rm -rf temp
+    fi
 }
 
 if [ ! -d "dependencies" ]; then
@@ -12,9 +16,9 @@ if [ ! -d "dependencies" ]; then
 fi
 
 
-install "dependencies/six-1.12.0.tar.gz"
-install "dependencies/pbr-5.2.0.tar.gz"
-install "dependencies/decorator-4.4.0.tar.gz"
-install "dependencies/ply-3.11.tar.gz"
-install "dependencies/jsonpath-rw-1.4.0.tar.gz"
-install "dependencies/jsonpath-rw-ext-1.2.0.tar.gz"
+install "dependencies/py_modules/six-1.12.0.tar.gz" "six"
+install "dependencies/py_modules/pbr-5.2.0.tar.gz" "pbr"
+install "dependencies/py_modules/decorator-4.4.0.tar.gz" "decorator"
+install "dependencies/py_modules/ply-3.11.tar.gz" "ply"
+install "dependencies/py_modules/jsonpath-rw-1.4.0.tar.gz" "jsonpath-rw"
+install "dependencies/py_modules/jsonpath-rw-ext-1.2.0.tar.gz" "jsonpath-rw-ext"
